requirements.txt
albucore==0.0.23
albumentations==2.0.4
annotated-types==0.7.0
anyio==4.8.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
contourpy==1.3.1
cycler==0.12.1
doclayout_yolo==0.0.3
fastapi==0.115.8
filelock==3.17.0
fonttools==4.56.0
fsspec==2025.2.0
gradio==5.16.0
gradio_client==1.7.0
h11==0.14.0
huggingface-hub==0.28.1
idna==3.10
Jinja2==3.1.5
kiwisolver==1.4.8
MarkupSafe==2.1.5
matplotlib==3.10.0
mpmath==1.3.0
networkx==3.4.2
numpy==2.2.2
opencv-python==4.11.0.86
opencv-python-headless==4.11.0.86
packaging==24.2
pandas==2.2.3
pillow==11.1.0
psutil==6.1.1
py-cpuinfo==9.0.0
pydantic==2.10.6
pydantic_core==2.27.2
PyMuPDF==1.27.1
pymupdf4llm==0.3.4
pymupdf_layout==1.27.1
pyparsing==3.2.1
python-dateutil==2.9.0.post0
python-multipart==0.0.20
python-dotenv==1.0.1
pytz==2025.1
PyYAML==6.0.2
requests==2.32.3
scipy==1.15.1
seaborn==0.13.2
setuptools==75.8.0
simsimd==6.2.1
six==1.17.0
sniffio==1.3.1
starlette==0.45.3
stringzilla==3.11.3
sympy==1.13.1
thop==0.1.1.post2209072238
torch==2.6.0
torchvision==0.21.0
tqdm==4.67.1
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
uvicorn==0.34.0
